requirements.txt
tensorflow-2.18.0
torch-2.5.1
transformers-4.46.3
numpy-1.26.4
keras-3.7.0
